refactor(indexer): split EventHandlers.ts monolith into focused modules
Break the 2,532-line EventHandlers.ts into 12 cohesive modules:

- helpers.ts — pure utilities (eventId, asAddress, asBigInt, hourBucket)
- abis.ts — ABI definitions (FPMM, SortedOracles, ERC20)
- priceDifference.ts — price math (computePriceDifference, normalizeTo18)
- tradingLimits.ts — trading limit types and computation
- rpc.ts — RPC client, fetch functions, caches, test mocks
- pool.ts — Pool/PoolSnapshot upsert, health status
- feeToken.ts — fee token metadata, backfill, YIELD_SPLIT_ADDRESS
- handlers/fpmm.ts — FPMMFactory + FPMM event handlers
- handlers/sortedOracles.ts — SortedOracles event handlers
- handlers/virtualPool.ts — VirtualPool event handlers
- handlers/feeToken.ts — ERC20FeeToken.Transfer (with eventFilters)

EventHandlers.ts becomes a slim entry point (52 lines) that imports
handler modules and re-exports test utilities for backwards compatibility.

All 73 existing tests pass without modification. No behavioral changes —
this is a pure structural refactoring.
