+++ indexer-envio/src/handlers/virtualPool.ts
@@ -1,0 +1,335 @@
+// ---------------------------------------------------------------------------
+// VirtualPoolFactory and VirtualPool event handlers
+// ---------------------------------------------------------------------------
+
+import {
+  VirtualPoolFactory,
+  VirtualPool,
+  type SwapEvent,
+  type LiquidityEvent,
+  type ReserveUpdate,
+  type RebalanceEvent,
+  type VirtualPoolLifecycle,
+} from "generated";
+import { eventId, asAddress, asBigInt } from "../helpers";
+import { fetchReserves } from "../rpc";
+import { upsertPool, upsertSnapshot, DEFAULT_ORACLE_FIELDS } from "../pool";
+
+// ---------------------------------------------------------------------------
+// VirtualPoolFactory.VirtualPoolDeployed
+// ---------------------------------------------------------------------------
+
+VirtualPoolFactory.VirtualPoolDeployed.handler(async ({ event, context }) => {
+  const id = eventId(event.chainId, event.block.number, event.logIndex);
+  const poolId = asAddress(event.params.pool);
+  const token0 = asAddress(event.params.token0);
+  const token1 = asAddress(event.params.token1);
+
+  await upsertPool({
+    context,
+    chainId: event.chainId,
+    poolId,
+    token0,
+    token1,
+    source: "virtual_pool_factory",
+    blockNumber: asBigInt(event.block.number),
+    blockTimestamp: asBigInt(event.block.timestamp),
+    oracleDelta: {
+      ...DEFAULT_ORACLE_FIELDS,
+      healthStatus: "N/A",
+    },
+  });
+
+  const lifecycle: VirtualPoolLifecycle = {
+    id,
+    poolId,
+    action: "DEPLOYED",
+    token0,
+    token1,
+    factoryAddress: asAddress(event.srcAddress),
+    txHash: event.transaction.hash,
+    blockNumber: asBigInt(event.block.number),
+    blockTimestamp: asBigInt(event.block.timestamp),
+  };
+
+  context.VirtualPoolLifecycle.set(lifecycle);
+});
+
+// ---------------------------------------------------------------------------
+// VirtualPoolFactory.PoolDeprecated
+// ---------------------------------------------------------------------------
+
+VirtualPoolFactory.PoolDeprecated.handler(async ({ event, context }) => {
+  const id = eventId(event.chainId, event.block.number, event.logIndex);
+  const poolId = asAddress(event.params.pool);
+
+  await upsertPool({
+    context,
+    chainId: event.chainId,
+    poolId,
+    source: "virtual_pool_factory",
+    blockNumber: asBigInt(event.block.number),
+    blockTimestamp: asBigInt(event.block.timestamp),
+  });
+
+  const lifecycle: VirtualPoolLifecycle = {
+    id,
+    poolId,
+    action: "DEPRECATED",
+    token0: undefined,
+    token1: undefined,
+    factoryAddress: asAddress(event.srcAddress),
+    txHash: event.transaction.hash,
+    blockNumber: asBigInt(event.block.number),
+    blockTimestamp: asBigInt(event.block.timestamp),
+  };
+
+  context.VirtualPoolLifecycle.set(lifecycle);
+});
+
+// ---------------------------------------------------------------------------
+// VirtualPool.Swap
+// ---------------------------------------------------------------------------
+
+VirtualPool.Swap.handler(async ({ event, context }) => {
+  const id = eventId(event.chainId, event.block.number, event.logIndex);
+  const poolId = asAddress(event.srcAddress);
+  const blockNumber = asBigInt(event.block.number);
+  const blockTimestamp = asBigInt(event.block.timestamp);
+
+  const volume0 =
+    event.params.amount0In > event.params.amount0Out
+      ? event.params.amount0In
+      : event.params.amount0Out;
+  const volume1 =
+    event.params.amount1In > event.params.amount1Out
+      ? event.params.amount1In
+      : event.params.amount1Out;
+
+  const reservesDelta = await fetchReserves(
+    event.chainId,
+    event.srcAddress,
+    blockNumber,
+  );
+
+  const pool = await upsertPool({
+    context,
+    chainId: event.chainId,
+    poolId,
+    source: "fpmm_swap", // reuse source key; VirtualPool inherits same priority
+    blockNumber,
+    blockTimestamp,
+    swapDelta: { volume0, volume1 },
+    ...(reservesDelta && { reservesDelta }),
+  });
+
+  await upsertSnapshot({
+    context,
+    pool,
+    blockTimestamp,
+    blockNumber,
+    swapDelta: { volume0, volume1 },
+  });
+
+  const swap: SwapEvent = {
+    id,
+    poolId,
+    sender: asAddress(event.params.sender),
+    recipient: asAddress(event.params.to),
+    amount0In: event.params.amount0In,
+    amount1In: event.params.amount1In,
+    amount0Out: event.params.amount0Out,
+    amount1Out: event.params.amount1Out,
+    txHash: event.transaction.hash,
+    blockNumber,
+    blockTimestamp,
+  };
+
+  context.SwapEvent.set(swap);
+});
+
+// ---------------------------------------------------------------------------
+// VirtualPool.Mint
+// ---------------------------------------------------------------------------
+
+VirtualPool.Mint.handler(async ({ event, context }) => {
+  const id = eventId(event.chainId, event.block.number, event.logIndex);
+  const poolId = asAddress(event.srcAddress);
+  const blockNumber = asBigInt(event.block.number);
+  const blockTimestamp = asBigInt(event.block.timestamp);
+
+  const pool = await upsertPool({
+    context,
+    chainId: event.chainId,
+    poolId,
+    source: "fpmm_mint",
+    blockNumber,
+    blockTimestamp,
+  });
+
+  await upsertSnapshot({
+    context,
+    pool,
+    blockTimestamp,
+    blockNumber,
+    mintDelta: true,
+  });
+
+  const liquidityEvent: LiquidityEvent = {
+    id,
+    poolId,
+    kind: "MINT",
+    sender: asAddress(event.params.sender),
+    recipient: asAddress(event.params.to),
+    amount0: event.params.amount0,
+    amount1: event.params.amount1,
+    liquidity: event.params.liquidity,
+    txHash: event.transaction.hash,
+    blockNumber,
+    blockTimestamp,
+  };
+
+  context.LiquidityEvent.set(liquidityEvent);
+});
+
+// ---------------------------------------------------------------------------
+// VirtualPool.Burn
+// ---------------------------------------------------------------------------
+
+VirtualPool.Burn.handler(async ({ event, context }) => {
+  const id = eventId(event.chainId, event.block.number, event.logIndex);
+  const poolId = asAddress(event.srcAddress);
+  const blockNumber = asBigInt(event.block.number);
+  const blockTimestamp = asBigInt(event.block.timestamp);
+
+  const pool = await upsertPool({
+    context,
+    chainId: event.chainId,
+    poolId,
+    source: "fpmm_burn",
+    blockNumber,
+    blockTimestamp,
+  });
+
+  await upsertSnapshot({
+    context,
+    pool,
+    blockTimestamp,
+    blockNumber,
+    burnDelta: true,
+  });
+
+  const liquidityEvent: LiquidityEvent = {
+    id,
+    poolId,
+    kind: "BURN",
+    sender: asAddress(event.params.sender),
+    recipient: asAddress(event.params.to),
+    amount0: event.params.amount0,
+    amount1: event.params.amount1,
+    liquidity: event.params.liquidity,
+    txHash: event.transaction.hash,
+    blockNumber,
+    blockTimestamp,
+  };
+
+  context.LiquidityEvent.set(liquidityEvent);
+});
+
+// ---------------------------------------------------------------------------
+// VirtualPool.UpdateReserves
+// ---------------------------------------------------------------------------
+
+VirtualPool.UpdateReserves.handler(async ({ event, context }) => {
+  const id = eventId(event.chainId, event.block.number, event.logIndex);
+  const poolId = asAddress(event.srcAddress);
+  const blockNumber = asBigInt(event.block.number);
+  const blockTimestamp = asBigInt(event.block.timestamp);
+
+  const pool = await upsertPool({
+    context,
+    chainId: event.chainId,
+    poolId,
+    source: "fpmm_update_reserves",
+    blockNumber,
+    blockTimestamp,
+    reservesDelta: {
+      reserve0: event.params.reserve0,
+      reserve1: event.params.reserve1,
+    },
+  });
+
+  await upsertSnapshot({
+    context,
+    pool,
+    blockTimestamp,
+    blockNumber,
+  });
+
+  const reserveUpdate: ReserveUpdate = {
+    id,
+    poolId,
+    reserve0: event.params.reserve0,
+    reserve1: event.params.reserve1,
+    blockTimestampInPool: event.params.blockTimestamp,
+    txHash: event.transaction.hash,
+    blockNumber,
+    blockTimestamp,
+  };
+
+  context.ReserveUpdate.set(reserveUpdate);
+});
+
+// ---------------------------------------------------------------------------
+// VirtualPool.Rebalanced
+// ---------------------------------------------------------------------------
+
+VirtualPool.Rebalanced.handler(async ({ event, context }) => {
+  // VirtualPools shouldn't normally rebalance, but handle defensively
+  const id = eventId(event.chainId, event.block.number, event.logIndex);
+  const poolId = asAddress(event.srcAddress);
+  const blockNumber = asBigInt(event.block.number);
+  const blockTimestamp = asBigInt(event.block.timestamp);
+
+  const pool = await upsertPool({
+    context,
+    chainId: event.chainId,
+    poolId,
+    source: "fpmm_rebalanced",
+    blockNumber,
+    blockTimestamp,
+    rebalanceDelta: true,
+  });
+
+  await upsertSnapshot({
+    context,
+    pool,
+    blockTimestamp,
+    blockNumber,
+    rebalanceDelta: true,
+  });
+
+  const priceDifferenceBefore = event.params.priceDifferenceBefore;
+  const priceDifferenceAfter = event.params.priceDifferenceAfter;
+  const improvement = priceDifferenceBefore - priceDifferenceAfter;
+  const effectivenessRatio =
+    priceDifferenceBefore > 0n
+      ? (Number(improvement) / Number(priceDifferenceBefore)).toFixed(4)
+      : "0.0000";
+
+  const rebalanced: RebalanceEvent = {
+    id,
+    poolId,
+    sender: asAddress(event.params.sender),
+    caller: event.transaction.from ?? "",
+    priceDifferenceBefore,
+    priceDifferenceAfter,
+    improvement,
+    effectivenessRatio,
+    txHash: event.transaction.hash,
+    blockNumber,
+    blockTimestamp,
+  };
+
+  context.RebalanceEvent.set(rebalanced);
+});
